Fixes routing issues and adds shop navigation
Adds a nav bar that lets users select from a list of clothing
categories which will then render all products matching the category
on the page through react-router-dom. Also fixes routing issues, namely
going to a new route through a Link component doesn't run the useEffect
method of the Products component. Hence, the component is rendered
but it does not render with the updated products. This fixes the problem
by adding the product category as a dependency.